+++ packages/firebase-dart/lib/src/auth.dart
@@ -24,6 +24,9 @@ class UserInfo<T extends firebase_interop.UserInfoJsImpl>
   /// User's e-mail address.
   String get email => jsObject.email;
 
+  /// The user's E.164 formatted phone number (if available).
+  String get phoneNumber => jsObject.phoneNumber;
+
   /// User's profile picture URL.
   String get photoURL => jsObject.photoURL;
 
@@ -91,6 +94,9 @@ class User extends UserInfo<firebase_interop.UserJsImpl> {
   ///
   /// Returns the current token if it has not expired, otherwise this will
   /// refresh the token and return a new one.
+  ///
+  /// It forces refresh regardless of token expiration if [forceRefresh]
+  /// parameter is [true].
   Future<String> getIdToken([bool forceRefresh = false]) =>
       handleThenable(jsObject.getIdToken(forceRefresh));
 
@@ -106,6 +112,15 @@ class User extends UserInfo<firebase_interop.UserJsImpl> {
   Future<User> linkWithCredential(AuthCredential credential) =>
       handleThenableWithMapper(
           jsObject.linkWithCredential(credential), User.get);
+
+  /// Links the user account with the given [phoneNumber] in E.164 format
+  /// (e.g. +16505550101) and [applicationVerifier].
+  Future<ConfirmationResult> linkWithPhoneNumber(
+          String phoneNumber, ApplicationVerifier applicationVerifier) =>
+      handleThenableWithMapper(
+          jsObject.linkWithPhoneNumber(
+              phoneNumber, applicationVerifier.jsObject),
+          (c) => new ConfirmationResult.fromJsObject(c));
 
   /// Links the authenticated [provider] to the user account using
   /// a pop-up based OAuth flow.
@@ -128,6 +143,18 @@ class User extends UserInfo<firebase_interop.UserJsImpl> {
       handleThenableWithMapper(
           jsObject.reauthenticateAndRetrieveDataWithCredential(credential),
           (o) => new UserCredential.fromJsObject(o));
+
+  /// Re-authenticates a user using a fresh credential.
+  /// Use before operations such as [updatePassword] that require tokens
+  /// from recent sign-in attempts.
+  ///
+  /// The user's phone number is in E.164 format (e.g. +16505550101).
+  Future<ConfirmationResult> reauthenticateWithPhoneNumber(
+          String phoneNumber, ApplicationVerifier applicationVerifier) =>
+      handleThenableWithMapper(
+          jsObject.reauthenticateWithPhoneNumber(
+              phoneNumber, applicationVerifier.jsObject),
+          (c) => new ConfirmationResult.fromJsObject(c));
 
   /// Re-authenticates a user using a fresh [credential]. Should be used
   /// before operations such as [updatePassword()] that require tokens
@@ -169,6 +196,10 @@ class User extends UserInfo<firebase_interop.UserJsImpl> {
   Future updatePassword(String newPassword) =>
       handleThenable(jsObject.updatePassword(newPassword));
 
+  /// Updates the user's phone number.
+  Future updatePhoneNumber(AuthCredential phoneCredential) =>
+      handleThenable(jsObject.updatePhoneNumber(phoneCredential));
+
   /// Updates a user's [profile] data.
   /// UserProfile has a displayName and photoURL.
   ///
@@ -201,6 +232,9 @@ class Auth extends JsObjectWrapper<AuthJsImpl> {
 
   /// Sends events when the users sign-in state changes.
   ///
+  /// After 4.0.0, this is only triggered on sign-in or sign-out.
+  /// To keep the old behavior, see [onIdTokenChanged].
+  ///
   /// If the value is `null`, there is no signed-in user.
   Stream<User> get onAuthStateChanged {
     if (_changeController == null) {
@@ -230,8 +264,12 @@ class Auth extends JsObjectWrapper<AuthJsImpl> {
   Func0 _onIdTokenChangedUnsubscribe;
   StreamController<User> _idTokenChangedController;
 
-  /// Sends events when the signed-in user's ID token, which includes sign-in,
-  /// sign-out, and token refresh events.
+  /// Sends events for changes to the signed-in user's ID token,
+  /// which includes sign-in, sign-out, and token refresh events.
+  ///
+  /// This method has the same behavior as [onAuthStateChanged] had prior to 4.0.0.
+  ///
+  /// If the value is `null`, there is no signed-in user.
   Stream<User> get onIdTokenChanged {
     if (_idTokenChangedController == null) {
       var nextWrapper = allowInterop((firebase_interop.UserJsImpl user) {
@@ -342,6 +380,23 @@ class Auth extends JsObjectWrapper<AuthJsImpl> {
   Future<User> signInWithEmailAndPassword(String email, String password) =>
       handleThenableWithMapper(
           jsObject.signInWithEmailAndPassword(email, password), User.get);
+
+  /// Asynchronously signs in using a phone number in E.164 format
+  /// (e.g. +16505550101).
+  ///
+  /// This method sends a code via SMS to the given phone number, and returns
+  /// a [ConfirmationResult].
+  /// After the user provides the code sent to their phone, call
+  /// [ConfirmationResult.confirm] with the code to sign the user in.
+  ///
+  /// For abuse prevention, this method also requires a [ApplicationVerifier].
+  /// The Firebase Auth SDK includes a reCAPTCHA-based implementation, [RecaptchaVerifier].
+  Future<ConfirmationResult> signInWithPhoneNumber(
+          String phoneNumber, ApplicationVerifier applicationVerifier) =>
+      handleThenableWithMapper(
+          jsObject.signInWithPhoneNumber(
+              phoneNumber, applicationVerifier.jsObject),
+          (c) => new ConfirmationResult.fromJsObject(c));
 
   /// Signs in using a popup-based OAuth authentication flow with the
   /// given [provider].
@@ -534,30 +589,139 @@ class TwitterAuthProvider extends AuthProvider<TwitterAuthProviderJsImpl> {
       TwitterAuthProviderJsImpl.credential(token, secret);
 }
 
-/// Phone auth provider.
+/// Phone number auth provider.
 ///
 /// See: <https://firebase.google.com/docs/reference/js/firebase.auth.PhoneAuthProvider>.
 class PhoneAuthProvider extends AuthProvider<PhoneAuthProviderJsImpl> {
   static String get PROVIDER_ID => PhoneAuthProviderJsImpl.PROVIDER_ID;
 
-  factory PhoneAuthProvider() =>
-      new PhoneAuthProvider.fromJsObject(new PhoneAuthProviderJsImpl());
+  /// Creates a new PhoneAuthProvider with the optional [Auth] instance
+  /// in which sign-ins should occur.
+  factory PhoneAuthProvider([Auth auth]) =>
+      new PhoneAuthProvider.fromJsObject((auth != null)
+          ? new PhoneAuthProviderJsImpl(auth.jsObject)
+          : new PhoneAuthProviderJsImpl());
 
+  /// Creates a new PhoneAuthProvider from a [jsObject].
   PhoneAuthProvider.fromJsObject(PhoneAuthProviderJsImpl jsObject)
       : super.fromJsObject(jsObject);
 
-  //(TODO)
-  // https://firebase.google.com/docs/reference/js/firebase.auth.PhoneAuthProvider#verifyPhoneNumber
-  // verifyPhoneNumber
+  /// Starts a phone number authentication flow by sending a verification code
+  /// to the given [phoneNumber] in E.164 format (e.g. +16505550101).
+  /// Returns an ID that can be passed to [PhoneAuthProvider.credential]
+  /// to identify this flow.
+  ///
+  /// For abuse prevention, this method also requires an [ApplicationVerifier].
+  Future<String> verifyPhoneNumber(
+          String phoneNumber, ApplicationVerifier applicationVerifier) =>
+      handleThenable(jsObject.verifyPhoneNumber(
+          phoneNumber, applicationVerifier.jsObject));
 
-  //(TODO)
-  // https://firebase.google.com/docs/reference/js/firebase.auth.PhoneAuthProvider#.credential
-  // credential
+  /// Creates a phone auth credential given the verification ID
+  /// from [verifyPhoneNumber] and the [verificationCode] that was sent to the
+  /// user's mobile device.
+  static AuthCredential credential(
+          String verificationId, String verificationCode) =>
+      PhoneAuthProviderJsImpl.credential(verificationId, verificationCode);
+}
+
+/// A verifier for domain verification and abuse prevention.
+///
+/// See: <https://firebase.google.com/docs/reference/js/firebase.auth.ApplicationVerifier>
+abstract class ApplicationVerifier<T extends ApplicationVerifierJsImpl>
+    extends JsObjectWrapper<T> {
+  /// Returns the type of application verifier (e.g. "recaptcha").
+  String get type => jsObject.type;
+
+  /// Creates a new ApplicationVerifier from a [jsObject].
+  ApplicationVerifier.fromJsObject(T jsObject) : super.fromJsObject(jsObject);
+
+  /// Executes the verification process.
+  /// Returns a Future containing string for a token that can be used to
+  /// assert the validity of a request.
+  Future<String> verify() => handleThenable(jsObject.verify());
+}
+
+/// reCAPTCHA verifier.
+///
+/// See: <https://firebase.google.com/docs/reference/js/firebase.auth.RecaptchaVerifier>
+/// See: <https://www.google.com/recaptcha/>
+class RecaptchaVerifier extends ApplicationVerifier<RecaptchaVerifierJsImpl> {
+  /// Creates a new RecaptchaVerifier from [container], [parameters] and [app].
+  ///
+  /// The [container] has different meaning depending on whether the reCAPTCHA
+  /// is hidden or visible. For a visible reCAPTCHA it must be empty.
+  /// If a string is used, it has to correspond to an element ID.
+  /// The corresponding element must also must be in the DOM at the time
+  /// of initialization.
+  ///
+  /// The [parameters] are optional [Map] of Recaptcha parameters.
+  /// See: <https://developers.google.com/recaptcha/docs/display#render_param>.
+  /// All parameters are accepted except for the sitekey.
+  /// Firebase Auth backend provisions a reCAPTCHA for each project
+  /// and will configure this upon rendering.
+  /// For an invisible reCAPTCHA, a size key must have the value 'invisible'.
+  ///
+  /// The [app] is the corresponding Firebase app.
+  /// If none is provided, the default Firebase App instance is used.
+  ///
+  ///     verifier = new fb.RecaptchaVerifier("register", {
+  ///       "size": "invisible",
+  ///       "callback": (resp) {
+  ///         print("Successful reCAPTCHA response");
+  ///       },
+  ///       "expired-callback": () {
+  ///         print("Response expired");
+  ///       }
+  ///     });
+  factory RecaptchaVerifier(container,
+          [Map<String, dynamic> parameters, App app]) =>
+      (parameters != null)
+          ? ((app != null)
+              ? new RecaptchaVerifier.fromJsObject(new RecaptchaVerifierJsImpl(
+                  container, jsify(parameters), app.jsObject))
+              : new RecaptchaVerifier.fromJsObject(
+                  new RecaptchaVerifierJsImpl(container, jsify(parameters))))
+          : new RecaptchaVerifier.fromJsObject(
+              new RecaptchaVerifierJsImpl(container));
+
+  /// Creates a new RecaptchaVerifier from a [jsObject].
+  RecaptchaVerifier.fromJsObject(RecaptchaVerifierJsImpl jsObject)
+      : super.fromJsObject(jsObject);
+
+  /// Clears the reCAPTCHA widget from the page and destroys the current instance.
+  clear() => jsObject.clear();
+
+  /// Renders the reCAPTCHA widget on the page.
+  /// Returns a Future that resolves with the reCAPTCHA widget ID.
+  Future<num> render() => handleThenable(jsObject.render());
+}
+
+/// A result from a phone number sign-in, link, or reauthenticate call.
+///
+/// See: <https://firebase.google.com/docs/reference/js/firebase.auth.ConfirmationResult>
+class ConfirmationResult extends JsObjectWrapper<ConfirmationResultJsImpl> {
+  /// Returns the phone number authentication operation's verification ID.
+  /// This can be used along with the verification code to initialize a phone
+  /// auth credential.
+  String get verificationId => jsObject.verificationId;
+
+  /// Creates a new ConfirmationResult from a [jsObject].
+  ConfirmationResult.fromJsObject(ConfirmationResultJsImpl jsObject)
+      : super.fromJsObject(jsObject);
+
+  /// Finishes a phone number sign-in, link, or reauthentication, given
+  /// the code that was sent to the user's mobile device.
+  Future<UserCredential> confirm(String verificationCode) =>
+      handleThenableWithMapper(jsObject.confirm(verificationCode),
+          (u) => new UserCredential.fromJsObject(u));
 }
 
 /// A structure containing a [User], an [AuthCredential] and [operationType].
 /// operationType could be 'signIn' for a sign-in operation, 'link' for a
 /// linking operation and 'reauthenticate' for a reauthentication operation.
+///
+/// See: <https://firebase.google.com/docs/reference/js/firebase.auth#.UserCredential>
 class UserCredential extends JsObjectWrapper<UserCredentialJsImpl> {
   /// Returns the user.
   User get user => User.get(jsObject.user);
@@ -568,7 +732,30 @@ class UserCredential extends JsObjectWrapper<UserCredentialJsImpl> {
   /// Returns the operation type.
   String get operationType => jsObject.operationType;
 
+  /// Returns additional user information from a federated identity provider.
+  AdditionalUserInfo get additionalUserInfo =>
+      new AdditionalUserInfo.fromJsObject(jsObject.additionalUserInfo);
+
   /// Creates a new UserCredential from a [jsObject].
   UserCredential.fromJsObject(UserCredentialJsImpl jsObject)
+      : super.fromJsObject(jsObject);
+}
+
+/// A structure containing additional user information from
+/// a federated identity provider.
+///
+/// See: <https://firebase.google.com/docs/reference/js/firebase.auth#.AdditionalUserInfo>
+class AdditionalUserInfo extends JsObjectWrapper<AdditionalUserInfoJsImpl> {
+  /// Returns the provider id.
+  String get providerId => jsObject.providerId;
+
+  /// Returns the profile.
+  Map<String, dynamic> get profile => dartify(jsObject.profile);
+
+  /// Returns the user name.
+  String get username => jsObject.username;
+
+  /// Creates a new AdditionalUserInfo from a [jsObject].
+  AdditionalUserInfo.fromJsObject(AdditionalUserInfoJsImpl jsObject)
       : super.fromJsObject(jsObject);
 }